requirements.txt.txt
streamlit==1.22.0
whois
dnspython
requests
